Use human-readable relative times in LoadGameActivity.
When loading a saved game, display the last time played in human
readable relative time. For example, "5 minutes ago" or "Yesterday".
This matches the template from list_entry_layout.xml

The current implementation writes the date in full, including day,
month, year, hours, minutes, and seconds. Not only is that more
information than we really need, but in devices with a narrower screen
this can cause the date to be drawn on top of the difficulty.